requirements.txt
httpx
pydantic-settings
tenacity
python-dotenv
supabase
